Fix TURN and STUN utils to work correctly under 64-bit Linux.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@10926 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXportLib/include/os/StunUtils.h
@@ -69,7 +69,7 @@ static bool allocateTurnAddress(OsDatagramSocket* pSocket,
                                 const int port,
                                 char* szRelayIp,
                                 unsigned short* piRelayPort,
-                                unsigned long* plLifetime) ;
+                                uint32_t* plLifetime) ;
 
 
 static const char* natTypeToString(STUN_NAT_TYPE type) ;
